connect: restore powered by, fixes #1088
