+++ wormhole-connect/src/views/Bridge/Send.tsx
@@ -1,7 +1,6 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Context } from '@wormhole-foundation/wormhole-connect-sdk';
-import { useTheme } from '@mui/material/styles';
 import { makeStyles } from 'tss-react/mui';
 
 import { CHAINS, TOKENS } from 'config';
@@ -31,7 +30,6 @@ import {
 import Button from 'components/Button';
 import CircularProgress from '@mui/material/CircularProgress';
 import AlertBanner from 'components/AlertBanner';
-import PoweredByIcon from 'icons/PoweredBy';
 import { isCosmWasmChain } from 'utils/cosmos';
 import { estimateClaimGas, estimateSendGas } from 'utils/gas';
 import { validateSolanaTokenAccount } from '../../utils/transferValidation';
@@ -39,13 +37,6 @@ import { validateSolanaTokenAccount } from '../../utils/transferValidation';
 const useStyles = makeStyles()((theme) => ({
   body: {
     width: '100%',
-  },
-  poweredBy: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    gap: '8px',
-    marginTop: '24px',
   },
   tosDisclaimer: {
     display: 'flex',
@@ -62,7 +53,6 @@ const useStyles = makeStyles()((theme) => ({
 
 function Send(props: { valid: boolean }) {
   const { classes } = useStyles();
-  const theme = useTheme();
   const dispatch = useDispatch();
   const wallets = useSelector((state: RootState) => state.wallet);
   const { sending, receiving } = wallets;
@@ -267,10 +257,6 @@ function Send(props: { valid: boolean }) {
           </Button>
         </>
       )}
-
-      <div className={classes.poweredBy}>
-        <PoweredByIcon color={theme.palette.text.primary} />
-      </div>
     </div>
   );
 }
